+++ pages/index.tsx
@@ -1,3 +1,3 @@
-import Blog from './Blog';
+import Home from './Home';
 
-export default Blog;
+export default Home;
